+++ mod.ts
@@ -6,12 +6,11 @@ export type {
   AccountInformation,
   BaseParameters,
   BaseResponse,
-  EngineMap,
   GetBySearchIdParameters,
   Location,
   Locations,
   LocationsApiParameters,
-} from "./src/types.ts";
+} from "./src/types.d.ts";
 export {
   getAccount,
   getHtml,
@@ -21,55 +20,4 @@ export {
   getLocations,
 } from "./src/serpapi.ts";
 
-export type { GoogleParameters } from "./src/engines/google.ts";
-export type { GoogleJobsParameters } from "./src/engines/google_jobs.ts";
-export type { GoogleJobsListingParameters } from "./src/engines/google_jobs_listing.ts";
-export type { GoogleReverseImageParameters } from "./src/engines/google_reverse_image.ts";
-export type { GoogleScholarProfilesParameters } from "./src/engines/google_scholar_profiles.ts";
-export type { GoogleScholarParameters } from "./src/engines/google_scholar.ts";
-export type { GoogleScholarCiteParameters } from "./src/engines/google_scholar_cite.ts";
-export type { GoogleScholarAuthorParameters } from "./src/engines/google_scholar_author.ts";
-export type { GoogleProductParameters } from "./src/engines/google_product.ts";
-export type { GoogleMapsParameters } from "./src/engines/google_maps.ts";
-export type { GoogleMapsPhotosParameters } from "./src/engines/google_maps_photos.ts";
-export type { GoogleMapsReviewsParameters } from "./src/engines/google_maps_reviews.ts";
-export type { GoogleEventsParameters } from "./src/engines/google_events.ts";
-export type { GoogleAutocompleteParameters } from "./src/engines/google_autocomplete.ts";
-export type { GoogleRelatedQuestionsParameters } from "./src/engines/google_related_questions.ts";
-export type { GoogleTrendsParameters } from "./src/engines/google_trends.ts";
-export type { GoogleTrendsAutocompleteParameters } from "./src/engines/google_trends_autocomplete.ts";
-export type { GoogleFinanceMarketsParameters } from "./src/engines/google_finance_markets.ts";
-export type { GoogleImmersiveProductParameters } from "./src/engines/google_immersive_product.ts";
-export type { BingParameters } from "./src/engines/bing.ts";
-export type { BingNewsParameters } from "./src/engines/bing_news.ts";
-export type { BingImagesParameters } from "./src/engines/bing_images.ts";
-export type { BaiduParameters } from "./src/engines/baidu.ts";
-export type { BaiduNewsParameters } from "./src/engines/baidu_news.ts";
-export type { YahooParameters } from "./src/engines/yahoo.ts";
-export type { YahooImagesParameters } from "./src/engines/yahoo_images.ts";
-export type { YahooVideosParameters } from "./src/engines/yahoo_videos.ts";
-export type { YahooShoppingParameters } from "./src/engines/yahoo_shopping.ts";
-export type { EbayParameters } from "./src/engines/ebay.ts";
-export type { YandexParameters } from "./src/engines/yandex.ts";
-export type { YandexImagesParameters } from "./src/engines/yandex_images.ts";
-export type { YandexVideosParameters } from "./src/engines/yandex_videos.ts";
-export type { YoutubeParameters } from "./src/engines/youtube.ts";
-export type { WalmartParameters } from "./src/engines/walmart.ts";
-export type { WalmartProductParameters } from "./src/engines/walmart_product.ts";
-export type { WalmartProductReviewsParameters } from "./src/engines/walmart_product_reviews.ts";
-export type { HomeDepotParameters } from "./src/engines/home_depot.ts";
-export type { HomeDepotProductParameters } from "./src/engines/home_depot_product.ts";
-export type { LinkedinParameters } from "./src/engines/linkedin.ts";
-export type { LinkedinProfileParameters } from "./src/engines/linkedin_profile.ts";
-export type { DuckduckgoParameters } from "./src/engines/duckduckgo.ts";
-export type { GooglePlayProductParameters } from "./src/engines/google_play_product.ts";
-export type { GooglePlayParameters } from "./src/engines/google_play.ts";
-export type { AppleAppStoreParameters } from "./src/engines/apple_app_store.ts";
-export type { AppleReviewsParameters } from "./src/engines/apple_reviews.ts";
-export type { AppleProductParameters } from "./src/engines/apple_product.ts";
-export type { NaverParameters } from "./src/engines/naver.ts";
-export type { GoogleLensParameters } from "./src/engines/google_lens.ts";
-export type { GoogleLocalServicesParameters } from "./src/engines/google_local_services.ts";
-export type { GoogleAboutThisResultParameters } from "./src/engines/google_about_this_result.ts";
-export type { YelpParameters } from "./src/engines/yelp.ts";
-export type { YelpReviewsParameters } from "./src/engines/yelp_reviews.ts";
+export * from "./src/engines/index.d.ts";