+++ src/engines/google_jobs_listing.ts
@@ -1,8 +1,0 @@
-export type GoogleJobsListingParameters = {
-  /**
-   * Search Query
-   * Parameter defines the `job_id` string which can be obtained from Google Jobs
-   * API.
-   */
-  q: string;
-};