+++ src/engines/google_scholar_cite.ts
@@ -1,9 +1,0 @@
-export type GoogleScholarCiteParameters = {
-  /**
-   * Search Result ID
-   * Parameter defines the ID of an individual Google Scholar organic search result.
-   * You can find the ID inside the `result_id` by using our [Google Scholar
-   * API](https://serpapi.com/google-scholar-api).
-   */
-  q: string;
-};