+++ src/engines/linkedin_profile.ts
@@ -1,8 +1,0 @@
-export type LinkedinProfileParameters = {
-  /**
-   * Profile ID
-   * LinkedIn profile ID. Can be found in profile url, e.g. profile id for
-   * https://www.linkedin.com/in/john-doe is john-doe
-   */
-  profile_id: string;
-};